[REFACTOR] #407: Car Controller 리팩토링(Swagger API 개선)
