+++ backend/src/main/java/autoever2/cartag/cars/CarService.java
@@ -1,5 +1,6 @@
-package autoever2.cartag.service;
+package autoever2.cartag.cars;
 
+import autoever2.cartag.cars.dto.*;
 import autoever2.cartag.domain.car.*;
 import autoever2.cartag.domain.color.InnerColorDto;
 import autoever2.cartag.domain.color.OuterColorDto;
@@ -10,7 +11,7 @@ import autoever2.cartag.domain.quote.QuoteDataDto;
 import autoever2.cartag.domain.quote.QuoteInfoDto;
 import autoever2.cartag.exception.EmptyDataException;
 import autoever2.cartag.exception.ErrorCode;
-import autoever2.cartag.repository.CarRepository;
+import autoever2.cartag.cars.CarRepository;
 import autoever2.cartag.repository.ColorRepository;
 import autoever2.cartag.repository.ModelRepository;
 import autoever2.cartag.repository.OptionRepository;
