+++ backend/src/main/java/autoever2/cartag/service/ModelService.java
@@ -3,7 +3,7 @@ package autoever2.cartag.service;
 import autoever2.cartag.domain.model.*;
 import autoever2.cartag.exception.EmptyDataException;
 import autoever2.cartag.exception.ErrorCode;
-import autoever2.cartag.repository.CarRepository;
+import autoever2.cartag.cars.CarRepository;
 import autoever2.cartag.repository.ModelRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
